refactor: move Cosmos subprocess utils from _base.py to _cosmos_utils.py
